KSS Parser for PHP
Adds a Parser object that accepts an array of files to
scan for KSS Blocks using the Comment Parser object. Any
comments found by Comment Parser are turned into Section
objects so the different parts of a comment can easily
be retrieved. Each modifier in the modifiers section in a
comment become Modifier objects with classname and description
properties.

These classes are based on the ruby version of KSS.