Adds Concept of Extenders to Modifiers
Modifiers can now be marked as being done through extension instead
of with a modifier class. This is done by adding @extender after
the modifier. You must also define the baseExtensionClass for
which a modifier is extending from by defining %baseExtensionClass
before the modifiers.

Allows for KSS Blocks to have a title marked with # as in Markdown.