Allows Markup to be defined in KSS Comment Blocks
Adds the option to specify your markup directly in a KSS Comment
Block by including a Markup: flag and then specifying the desired
markup.

Markup is automatically passed to the modifiers and becomes
the default for the $modifier->getExampleHtml(). However, you
can still pass an html string to getExampleHtml() and that html
will be used instead of the default markup specified in the kss
comment block.

Updates test suite to include tests to check markup text handling.

Updates example app to use the embedded markup style of supplying
example html for the styleguide.

Updates readme with new markup information and also provides an
example of including a title in your styleguides.