Adds addition methods to getSections
Adds a method to return all sections found by the parser.

Adds a method to get all the top level sections from the parser.
Top level sections are defined as being either in the format of
1, 2, etc. or 1.0, 2.0, etc.

Adds a method to return all children of a specified section reference.

Adds a method to sort sections by depth and by section order.

Adds methods to Section to calculate their section depth and to
calculate their depth score that assists in ordering sections by
their depth.