Adds get, trim, and belongsToReference to Section
Adds a getReference to Section and chages getSection to call getReference.
Marks getSection as deprecated.

Adds trimReference method to Section that removes any trailing periods
and trims all placeholder zeroes in a section reference. In other words,
3.0.0 is considered to be the same as 3. 3.0.1, of course, remains as
3.0.1.

Adds a method to check if a section is a reference or a child of that
reference. For example, section 2.1.1 would be considered to belong to
reference 2, 2.1, and 2.1.1.

Parser->getSection now throws an UnexpectedValueException if a section
can not be found for the requested reference.

Adds corresponding tests.