Adds files with contributing guidelines
Adds a CONTRIBUTING.md file to show guidelines for contributors. Made
mention that code must follow the PSR guidelines and that test cases
must be added for new methods or bug fixes. Requests git commit messages
to follow the subject line and body style.

Adds an .editorconfig with whitespace defaults for IDE's that support
them.