Fixes output of description and extender modifiers
block.inc.php was incorrectly calling $modifier->getDescription()
resulting in no descriptions being shown for modifiers. Fixes call
to getDescription.

The extender modifier classes were being checked to see if they needed
to be added in two if statements but the classes were not being correctly
echoed. Fixes it so the classes are now probably echoed.