Adds Non-numeric Styleguide Reference Support
Updates the Section and Parser classes to support non-numeric Styleguide
references. References can now either be in the format of "Buttons.Star"
or "Buttons - Star" as well as the normal "1.2.1", etc.

Updates example project to user non-numeric Styleguide references with
dash delimiters. Updates README file in a similar manner.

Updates to latest version of PHPUnit 4.