Styleguide References are Matched At Start Of Line
To prevent matching the word "styleguide" that may be found in other
parts of the KSS Block, the Styleguide Reference is now required to
either be at the start of the line or immediately following a "//" at
the start of a line. Any number of spaces can proceed either.

Adds related tests and a test to make sure "No styleguide reference"
isn't treated as a valid reference.